% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_state_score.R
\name{average_state_score}
\alias{average_state_score}
\title{Average score of the tweets}
\usage{
average_state_score(downloaded_tweets)
}
\arguments{
\item{downloaded_tweets:}{A list of tweets and their characteristics
corresponding to the output of the function tweets_analysis}
}
\value{
A matrix of 49 rows and 2 columns, with the first row containing the
name of the state and the second the corresponding average happiness score of
the tweets published in that state
}
\description{
This function averages the happiness score of the tweets fetched
by the tweets_analysis function and stores it in a matrix with the name of
each state and their corresponding average score
}
\author{
group2
}
