% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending.R
\name{trending}
\alias{trending}
\title{Get Twitter's trending topics}
\usage{
trending(place)
}
\arguments{
\item{place:}{An integer corresponding to the WOEID (Yahoo! Where On Earth ID)
of the place you want the trending topics of. To get these WOEID, use the
function trends_available() from the rtweet package. This will retrun you a
table with the cities you can look up and their corresponding WOEID.}
}
\value{
A vector of variable length, containing the trending topics ordered by
decreasing popularity
}
\description{
This function allows the user to fetch the curernt trending
topics on Twitter in 63 cities of the United States or in the whole country by
specifying the city or by writing United States as input.
}
\author{
group2
}
