% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic api and NPL.R
\name{tweets_analysis}
\alias{tweets_analysis}
\title{Tweets Analysis}
\usage{
tweets_analysis()
}
\value{
A table containing the user name of the twitter user, the text, the
number of 'favorites' given to the tweet and the device used to send it
}
\description{
Analyses tweets in every state of the US
}
\author{
group2
}
