% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_analysis.R
\name{tweets_analysis}
\alias{tweets_analysis}
\title{Tweets Analysis}
\usage{
tweets_analysis(Hashtag = NULL, ...)
}
\arguments{
\item{Hashtag:}{A single string starting with a # and which must correspond
to an existing # on Twitter}

\item{...:}{Any additional Hastags the user wants to see. The function will
look for tweets having all the coresponding hashtags mentionned by the user,
not any of them}
}
\value{
A table containing the user name of the twitter user, the text, the
number of 'favorites' given to the tweet and the device used to send it
}
\description{
Analyses tweets in every state of the US
}
\author{
group2
}
